+++ packages/maskbook/src/plugins/Trader/apis/trending/hotfix.ts
@@ -52,18 +52,20 @@ export function resolveAlias(keyword: string, dataProvider: DataProvider) {
     if (dataProvider === DataProvider.COIN_GECKO)
         return KEYWORD_ALIAS_MAP[DataProvider.COIN_GECKO][keyword.toUpperCase()] ?? keyword
     if (dataProvider === DataProvider.UNISWAP) return keyword
-    unreachable(dataProvider as never)
+    unreachable(dataProvider)
 }
 
 export function resolveCoinId(keyword: string, dataProvider: DataProvider) {
     if (dataProvider === DataProvider.COIN_MARKET_CAP)
         return KEYWORK_ID_MAP[DataProvider.COIN_MARKET_CAP][keyword.toUpperCase()]
     if (dataProvider === DataProvider.COIN_GECKO) return KEYWORK_ID_MAP[DataProvider.COIN_GECKO][keyword.toUpperCase()]
-    unreachable(dataProvider as never)
+    if (dataProvider === DataProvider.UNISWAP) throw new Error('not implemented')
+    unreachable(dataProvider)
 }
 
 export function resolveCoinAddress(id: string, dataProvider: DataProvider) {
     if (dataProvider === DataProvider.COIN_MARKET_CAP) return ID_ADDRESS_MAP[DataProvider.COIN_MARKET_CAP][id]
     if (dataProvider === DataProvider.COIN_GECKO) return ID_ADDRESS_MAP[DataProvider.COIN_GECKO][id]
-    unreachable(dataProvider as never)
+    if (dataProvider === DataProvider.UNISWAP) throw new Error('not implemented')
+    unreachable(dataProvider)
 }
