+++ packages/maskbook/src/plugins/Trader/apis/uniswap/prediction_market_tokens.json
